refactor: move sliders creation into separate file
